+++ src/Components/Charts/TokenChart.vue
@@ -2,9 +2,9 @@
   <div class="TokenChart" ref="container" >
     <div class="time-range">
       Showing data from
-      <span class="date">{{ focusExtent[0].toDateString() }}</span>
+      <span class="date">{{ fromDate }}</span>
       to
-      <span class="date">{{ focusExtent[1].toDateString() }}</span>
+      <span class="date">{{ toDate }}</span>
     </div>
 
     <svg :id="plotId" >
@@ -97,10 +97,10 @@
 }
 
 .TokenChart .y-axis .tick line {
-  stroke: #F5F5F5;
+  stroke: #EAEAEA;
 }
 .TokenChart .context-y-axis .tick line {
-  stroke: #F5F5F5;
+  stroke: #EFEFEF;
 }
 
 .TokenChart .context-y-axis .tick text,
@@ -136,6 +136,7 @@
 
 <script>
 import * as d3 from 'd3';
+import * as moment from 'moment';
 
 
 const DAY_IN_MS = 24 * 60 * 60 * 1000;
@@ -143,7 +144,7 @@ const NOW = new Date();
 const TODAY = new Date(Date.UTC(
   NOW.getFullYear(),
   NOW.getMonth(),
-  NOW.getDate() + 1,
+  NOW.getDate(),
   0,
   0,
   0,
@@ -153,6 +154,7 @@ const ONE_MONTH_AGO = new Date(TODAY);
 
 ONE_WEEK_AGO.setDate(TODAY.getDate() - 6);
 ONE_MONTH_AGO.setMonth(TODAY.getMonth() - 1);
+
 
 /**
  * TokenChart component.
@@ -194,6 +196,15 @@ export default {
       bottom: 0,
     },
   }),
+
+  computed: {
+    fromDate() {
+      return moment.utc(this.focusExtent[0]).format('ddd, D MMM YYYY');
+    },
+    toDate() {
+      return moment.utc(this.focusExtent[1]).format('ddd, D MMM YYYY');
+    },
+  },
 
   /**
    * Method to be executed after the component has been mounted.
@@ -290,8 +301,19 @@ export default {
       let value;
 
       for (let i = 0; i < this.data.length; i++) {
-        positive = 0;
-        negative = 0;
+        // Both counts start in 1 and -1 to allow the maximum and minimum token 
+        // values to be included in the chart axis labels.
+        //
+        // This is due to d3.js optimizing the number of ticks/labels for the
+        // available space. d3 tries to include all the labels if there's enough
+        // space. If there is not, it will reduce the tick/label count to a
+        // dividend of the original count. Hence, if the maximum value is 15,
+        // the scale will go from 1 to 14 in steps of 2 (0, 2, 4, 6 and so on).
+        // This causes the 15 not to be included in the scale. Adding one to
+        // both ends makes sure the scale will go up to 16, which feels more
+        // natural.
+        positive = 1; 
+        negative = -1;
         dataPoint = this.data[i];
 
         for (let j = 0; j < this.features.length; j++) {
@@ -395,36 +417,33 @@ export default {
         .domain([ONE_MONTH_AGO, TODAY])
         .range([0, this.width + focusBarWidth]);
 
-      // Axes.
       const range = this.focusExtent;
       const timeDelta = range[1].getTime() - range[0].getTime();
       const numDays = Math.ceil(timeDelta / (24 * 60 * 60 * 1000));
+
+      // X-axis.
       const xAxis = d3
         .axisBottom()
         .scale(this.x)
         .ticks(d3.utcDay)
-        .tickSize(this.focusHeight)
-        .tickFormat(d => d.toLocaleDateString(
-          'default',
-          { day: 'numeric', month: 'short' },
-        ));
+        .tickSize(this.focusHeight)  // Height of the tick line.
+        .tickFormat(d => moment.utc(d).format('MMM D'));
       const contextXAxis = d3
         .axisBottom()
         .scale(this.contextX)
         .ticks(15)
-        .tickSize(this.contextHeight)
-        .tickFormat(d => d.toLocaleDateString(
-          'default',
-          { day: 'numeric', month: 'numeric' },
-        ));
+        .tickSize(this.contextHeight)  // Height of the tick line.
+        .tickFormat(d => moment.utc(d).format('M/D'));
       const yAxis = d3
         .axisLeft()
         .scale(this.y)
-        .tickSize(-this.width - focusBarWidth);
+        .tickSize(-this.width - focusBarWidth)  // Width of the tick line.
+        .tickFormat(d3.format('d'));
       const contextYAxis = d3
         .axisLeft()
         .scale(this.contextY)
-        .tickSize(-this.width - focusBarWidth);
+        .tickSize(-this.width - focusBarWidth) // Width of the tick line.
+        .tickFormat(d3.format('d'));
 
       // Append the axes.
       this.svg